Refactor router into separate controllers
